// frontend/src/sections/clinicians/clinicians.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type CliniciansVariables = {
  storeId: string;
  filter?: {
    id?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    code?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    firstName?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    lastName?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    initials?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    address1?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    address2?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    phone?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    mobile?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    email?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    isActive?: boolean | null;
  } | null;
  sort?: Array<{
    key: "code" | "firstName" | "lastName" | "initials" | "address1" | "address2" | "phone" | "mobile" | "email";
    desc?: boolean | null;
  }> | null;
  page?: {
    first?: number | null;
    offset?: number | null;
  } | null;
};

export type CliniciansResult = {
  clinicians: ({
  __typename: "ClinicianConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  code: string;
  firstName: string | null;
  lastName: string;
  initials: string;
  mobile: string | null;
  gender: "FEMALE" | "MALE" | "TRANSGENDER" | "TRANSGENDER_MALE" | "TRANSGENDER_MALE_HORMONE" | "TRANSGENDER_MALE_SURGICAL" | "TRANSGENDER_FEMALE" | "TRANSGENDER_FEMALE_HORMONE" | "TRANSGENDER_FEMALE_SURGICAL" | "UNKNOWN" | "NON_BINARY" | null;
}>;
});
};

export const Clinicians = {
  query: "query clinicians($storeId: String!, $filter: ClinicianFilterInput, $sort: [ClinicianSortInput!], $page: PaginationInput) {\n  clinicians(storeId: $storeId, filter: $filter, sort: $sort, page: $page) {\n    ... on ClinicianConnector {\n      __typename\n      totalCount\n      nodes {\n        id\n        code\n        firstName\n        lastName\n        initials\n        mobile\n        gender\n      }\n    }\n  }\n}",
} as TypedDocument<CliniciansResult, CliniciansVariables>;